Enable smart suggestions and completions in Visual Studio Code JavaScript projects
The language service of my Visual Studio Code ignores the node_modules and does not pick .d.ts-es from the tns-core-modules. The reference line enables the smart code completions and suggestions in all .js files within the app. It may become redundant in future versions of Visual Studio Code but for now it brings in a lot of value along the .d.ts-es we will ship with v1.5.0.